Polished code for Priority junction . Ignore arrows for one way main road.
